+++ RELEASE_NOTES.md
@@ -13,6 +13,14 @@ Release notes template:
 ## Removed
 
 -->
+
+# 2019-09-27
+
+## Added
+
+* Adding HathiTrust submission-information-package creator to exporters.
+
+
 # 2019-09-26
 
 ## Fixed